[app-emulation/qemu-launcher] new, requested in bug 3235
One can think of adding || app-emulation/qemu-kvm. However it doesn't
work with qemu-i386 with which I tested, but could work with
qemu-system-i386 but I can't test it (KVM). If anyone can and confirm
it does (need to specify correct executable in configuration first),
it would be nice to update the runtime dependency.