[sci-geosciences/qgis] Update ebuild for bug# 4928
The qgis ebuild has been failing when the spatialite USE flag is
not set.

https://bugs.sabayon.org/show_bug.cgi?id=4928

Upstream research indicates that spatialite seems to be a mandatory
dependency now.

http://hub.qgis.org/issues/8797

The ebuild in the rion overlay already reflects these changes, so I've
updated the qgis ebuild to version 2.8.1 (from the rion overlay) and
removed the spatialite USE flag, leaving spatialite as a hard
dependency now.

Old ebuilds and patches were cleaned out.

Signed-off-by: Mitch Harder <mitch.harder@sabayonlinux.org>